Commits missing on successor branches feature.
To detect unmerged (bug) fixes, we regularly need to analyze differences
between a 'released' branch and the current upstream branch.

This feature collects all commits that are not merged into the upstream
branch from an other branch.

The current implementation will report commits that have been
cherry-picked, but not merged.